Added more flexible creation
The othello board can now be created in three different ways.
Before this change, it was only possible to create a new board with the
standard othello opening. It is now also possible to create an entirely
empty board as well as provide the position of all the stones manually,
provided that no two stones overlap.

Signed-off-by: Emil Englesson <englesson.emil@gmail.com>