Fixed unsound deserialization bug
The OthelloBoard struct is careful to validate the board representation
to ensure that no stones overlap with each other. Unfortunately, this
validation was not performed when deserializing with Serde. The solution
was to introduce a shadow type which allowed us to avoid writing a
custom deserializer and instead reusing validation functions.

Signed-off-by: Emil Englesson <englesson.emil@gmail.com>