Abstracted u64 as Bitboard- and Position-structs
Instead of working with raw `u64`, Bitboard- and Position-structs were
implemented. The bitboard will be a fairly generic container but it will
allow the extension traits to be scrapped in favor of functions on the
Bitboard struct. The Position struct will represent a single position on
the board which will enable some functions to elide some errorchecking.
Furthermore, it has the ability to convert to and from various other
representations such as rank/file and algebraic notation.